PAE-1412: rename enrich-boom to cdp-boom and fold types
