--- src/common/types/enriched-boom.js
@@ -1,12 +1,0 @@
-/**
- * @import { Boom } from '@hapi/boom'
- *
- * @typedef {object} BoomEvent
- * @property {string} action
- * @property {string} reason
- * @property {string} [reference]
- *
- * @typedef {Boom & { code?: string, event?: BoomEvent }} EnrichedBoom
- */
-
-export {}
